+++ uniflow-android/src/main/java/io/uniflow/android/flow/AndroidDataFlow.kt
@@ -18,72 +18,73 @@ package io.uniflow.android.flow
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.MutableLiveData
 import android.arch.lifecycle.ViewModel
+import android.support.annotation.CallSuper
 import io.uniflow.core.dispatcher.UniFlowDispatcher
 import io.uniflow.core.flow.*
-import io.uniflow.core.logger.UniFlowLogger
+import io.uniflow.core.flow.data.Event
+import io.uniflow.core.flow.data.UIEvent
+import io.uniflow.core.flow.data.UIState
 import io.uniflow.core.threading.onMain
 import kotlinx.coroutines.*
-import kotlinx.coroutines.channels.actor
+import kotlinx.coroutines.channels.Channel
 
-abstract class AndroidDataFlow(defaultCapacity: Int = 10, override val defaultDispatcher: CoroutineDispatcher = UniFlowDispatcher.dispatcher.io()) : ViewModel(), DataFlow {
+/**
+ * Android implementation of [DataFlow].
+ * This is also a [ViewModel].
+ * Its [coroutineScope] uses [Dispatchers.Main] and is automatically cancelled if the `ViewModel`
+ * is cleared.
+ *
+ * @param defaultCapacity
+ * The default capacity of this `DataFlow`.
+ * Any state actions dispatched using [setState] will be added to the buffer unless it's full.
+ * Defaults to [Channel.BUFFERED].
+ *
+ * @param defaultDispatcher The default [CoroutineDispatcher] on which state actions are dispatched.
+ * Defaults to [Dispatchers.IO].
+ */
+abstract class AndroidDataFlow(
+    defaultState: UIState = UIState.Empty,
+    defaultCapacity: Int = Channel.BUFFERED,
+    defaultDispatcher: CoroutineDispatcher = UniFlowDispatcher.dispatcher.io()
+) : ViewModel(),
+    DataFlow,
+    UIDataPublisher {
 
-    private val viewModelJob = SupervisorJob()
-    override val coroutineScope: CoroutineScope = CoroutineScope(Dispatchers.Main + viewModelJob)
+    private val supervisorJob = SupervisorJob()
+    final override val coroutineScope = CoroutineScope(Dispatchers.Main + supervisorJob)
+
+    private val uiDataManager by lazy { UIDataManager(this, defaultState) }
+    final override val scheduler =
+        ActionFlowScheduler(uiDataManager, coroutineScope, defaultDispatcher, defaultCapacity)
 
     private val _states = MutableLiveData<UIState>()
-    val states: LiveData<UIState>
-        get() = _states
+    val states: LiveData<UIState> = _states
 
     private val _events = MutableLiveData<Event<UIEvent>>()
-    val events: LiveData<Event<UIEvent>>
-        get() = _events
+    val events: LiveData<Event<UIEvent>> = _events
 
-    override suspend fun sendEvent(event: UIEvent): UIState? {
-        onMain(immediate = true) {
-            UniFlowLogger.logEvent(event)
-            _events.value = Event(event)
-        }
-        return null
+    init {
+        action { setState { defaultState } }
     }
 
-    override suspend fun notifyUpdate(newState: UIState, notificationEvent: UIEvent): UIState? {
-        onMain(immediate = true) {
-            UniFlowLogger.logState(newState)
-            _internalState = newState
-            UniFlowLogger.logEvent(notificationEvent)
-            _events.value = Event(notificationEvent)
-        }
-        return null
-    }
+    final override fun getCurrentState() = uiDataManager.currentState
 
-    override suspend fun applyState(state: UIState) {
+    final override suspend fun publishState(state: UIState) {
         onMain(immediate = true) {
-            UniFlowLogger.logState(state)
-            _internalState = state
             _states.value = state
         }
     }
 
-    private var _internalState: UIState? = null
-
-    override val currentState: UIState?
-        get() = _internalState
-
-    override val actorFlow = coroutineScope.actor<StateAction>(UniFlowDispatcher.dispatcher.default(), capacity = defaultCapacity) {
-        for (action in channel) {
-            if (coroutineScope.isActive) {
-                withContext(defaultDispatcher) {
-                    proceedAction(action)
-                }
-            } else {
-                UniFlowLogger.log("actor $action cancelled")
-            }
+    final override suspend fun sendEvent(event: UIEvent) {
+        onMain(immediate = true) {
+            _events.value = Event(event)
         }
     }
 
+    @CallSuper
     override fun onCleared() {
+        coroutineScope.cancel()
+        scheduler.close()
         super.onCleared()
-        viewModelJob.cancel()
-        actorFlow.close()
     }
 }