+++ uniflow-android/src/main/java/io/uniflow/android/flow/DataFlowObserver.kt
@@ -17,9 +17,8 @@ package io.uniflow.android.flow
 
 import android.arch.lifecycle.LifecycleOwner
 import android.arch.lifecycle.Observer
-import io.uniflow.core.flow.Event
-import io.uniflow.core.flow.UIEvent
-import io.uniflow.core.flow.UIState
+import io.uniflow.core.flow.data.Event
+import io.uniflow.core.flow.data.UIState
 
 /**
  * AndroidDataFlow Observers for states & events