+++ uniflow-core/src/main/kotlin/io/uniflow/core/flow/data/UIData.kt
@@ -1,0 +1,3 @@
+package io.uniflow.core.flow.data
+
+interface UIData