+++ uniflow-test/src/test/kotlin/io/uniflow/test/StackFlowTest.kt
@@ -1,14 +1,14 @@
 package io.uniflow.test
 
-import io.uniflow.core.flow.UIEvent
-import io.uniflow.core.flow.UIState
+import io.uniflow.core.flow.data.UIEvent
+import io.uniflow.core.flow.data.UIState
 import io.uniflow.core.logger.DebugMessageLogger
 import io.uniflow.core.logger.UniFlowLogger
 import io.uniflow.core.logger.UniFlowLoggerTestRule
 import io.uniflow.test.data.Todo
 import io.uniflow.test.data.TodoListState
 import io.uniflow.test.data.TodoRepository
-import io.uniflow.test.data.TodoStackActorFlow
+import io.uniflow.test.impl.SampleFlow
 import io.uniflow.test.rule.TestDispatchersRule
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.runBlocking
@@ -34,11 +34,11 @@ class StackFlowTest {
     var rule = TestDispatchersRule()
 
     val repository = TodoRepository()
-    lateinit var dataFlow: TodoStackActorFlow
+    lateinit var dataFlow: SampleFlow
 
     @Before
     fun before() {
-        dataFlow = TodoStackActorFlow(repository)
+        dataFlow = SampleFlow(repository)
     }
 
     @Test
@@ -126,21 +126,6 @@ class StackFlowTest {
         assertTrue(dataFlow.events.size == 1)
     }
 
-    @Test
-    fun `action state error`() {
-        dataFlow.getAll()
-        dataFlow.add("first")
-        dataFlow.makeOnStateError()
-
-        assertEquals(UIState.Empty, dataFlow.states[0])
-        assertEquals(TodoListState(emptyList()), dataFlow.states[1])
-        assertEquals(TodoListState(listOf(Todo("first"))), dataFlow.states[2])
-
-        assertTrue(dataFlow.states.size == 3)
-        assertTrue(dataFlow.events.last() is UIEvent.Fail)
-        assertTrue(dataFlow.events.size == 1)
-    }
-
 //    @Test
 //    fun `action failed error`() {
 //        dataFlow.getAll()
@@ -202,7 +187,7 @@ class StackFlowTest {
         dataFlow.getAll()
         dataFlow.longWait()
         delay(300)
-        dataFlow.cancel()
+        dataFlow.close()
 
         assertEquals(UIState.Empty, dataFlow.states[0])
         assertEquals(TodoListState(emptyList()), dataFlow.states[1])
@@ -214,7 +199,7 @@ class StackFlowTest {
     @Test
     fun `cancel before test`() = runBlocking {
         dataFlow.getAll()
-        dataFlow.cancel()
+        dataFlow.close()
         dataFlow.longWait()
 
         assertEquals(UIState.Empty, dataFlow.states[0])